# api-gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.AccessTokenData:
    properties:
      expiresIn:
        type: integer
      token:
        type: string
    type: object
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.AccessTokenResponse:
    properties:
      accessToken:
        $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.AccessTokenData'
    type: object
...
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.CheckActivationCodeRequest:
    properties:
      code:
        type: string
    type: object
...
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorDetail:
    properties:
      message:
        type: string
    type: object
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorDetail'
    type: object
...
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.IDResponse:
    properties:
      id:
        type: string
    type: object
...
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
      phoneNumber:
        type: string
    type: object
...
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.RegisterRequest:
    properties:
      birthDate:
        type: string
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        type: string
      passwordConfirmation:
        type: string
      phoneNumber:
        type: string
    type: object
...
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.User:
    properties:
      birthDate:
        type: string
      createdAt:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: string
      isDocumentVerified:
        type: boolean
      isEmailVerified:
        type: boolean
      isSuspended:
        type: boolean
      lastName:
        type: string
      passwordHash:
        items:
          type: integer
        type: array
      phoneNumber:
        type: string
      profileImageURL:
        type: string
      roles:
        items:
          type: string
        type: array
      updatedAt:
        type: string
    type: object
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.UserCreateRequest:
    properties:
      birthDate:
        type: string
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        type: string
      passwordConfirmation:
        type: string
      phoneNumber:
        type: string
    type: object
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.UserProfileUpdateRequest:
    properties:
      birthDate:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        type: string
      passwordConfirmation:
        type: string
      phoneNumber:
        type: string
      profileImageKey:
        type: string
    type: object
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.UserResponse:
    properties:
      user:
        $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.User'
    type: object
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.UserUpdateRequest:
    properties:
      birthDate:
        type: string
      email:
        type: string
      firstName:
        type: string
      isDocumentVerified:
        type: boolean
      isEmailVerified:
        type: boolean
      isSuspended:
        type: boolean
      lastName:
        type: string
      password:
        type: string
      passwordConfirmation:
        type: string
      phoneNumber:
        type: string
      profileImageKey:
        type: string
      roles:
        items:
          type: string
        type: array
    type: object
  github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.UsersResponse:
    properties:
      users:
        items:
          $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.User'
        type: array
    type: object
...
host: localhost:4000
info:
  contact:
    email: support@carrental.kz
    name: API Support
  description: API Gateway for the Kazakhstan carsharing platform. Handles auth, users,
    fleet, zones, insurance, and maintenance.
  license:
    name: Proprietary
  termsOfService: http://swagger.io/terms/
  title: Car Rental API Gateway
  version: "1.0"
paths:
  /auth/refresh-token:
    post:
      description: Issues a new access token using the refresh token stored in the
        HttpOnly cookie.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.AccessTokenResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account. Returns the new user's ID.
      parameters:
      - description: Registration payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.IDResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: Authenticates a user by email or phone + password. Returns an access
        token and sets an HttpOnly refresh token cookie.
      parameters:
      - description: Sign-in credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.AccessTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
      summary: Sign in
      tags:
      - auth
  /auth/sign-out:
    post:
      description: Invalidates the current session and clears the refresh token cookie.
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Sign out
      tags:
      - auth
...
  /users:
    get:
      description: Returns a list of users, optionally filtered by query params.
      parameters:
      - description: Filter by email
        in: query
        name: email
        type: string
      - description: Filter by phone number
        in: query
        name: phoneNumber
        type: string
      - description: Filter by first name
        in: query
        name: firstName
        type: string
      - description: Filter by last name
        in: query
        name: lastName
        type: string
      - description: Filter by document verification status
        in: query
        name: isDocumentVerified
        type: boolean
      - description: Filter by email verification status
        in: query
        name: isEmailVerified
        type: boolean
      - description: Filter by suspension status
        in: query
        name: isSuspended
        type: boolean
      - description: Pagination limit
        in: query
        name: limit
        type: integer
      - description: Pagination offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.UsersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get users with filter
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Admin endpoint to create a user with explicit roles and activation
        state.
      parameters:
      - description: Create payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.UserCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.IDResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create user (admin)
      tags:
      - users
  /users/{id}:
    delete:
      description: Deletes a user by ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - users
    get:
      description: Returns a single user by their ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Partially updates a user by ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
...
  /users/activation-code/check:
    post:
      consumes:
      - application/json
      description: Verifies the activation code submitted by the user.
      parameters:
      - description: Activation code
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.CheckActivationCodeRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Check activation code
      tags:
      - users
  /users/activation-code/send:
    post:
      description: Sends an activation code to the current user's email or phone.
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Send activation code
      tags:
      - users
...
  /users/profile:
    get:
      description: Returns the profile of the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Partially updates the authenticated user's own profile. Critical
        fields (roles, verification flags) are not accepted.
      parameters:
      - description: Profile fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.UserProfileUpdateRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_sorawaslocked_car-rental-api-gateway_internal_adapter_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update current user profile
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer <access_token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"